+++ models/public_file.go
@@ -153,7 +153,7 @@ type PublicFileAttributes struct {
 	Provider string `json:"provider,omitempty"`
 
 	// size
-	Size float64 `json:"size,omitempty"`
+	Size int64 `json:"size,omitempty"`
 
 	// thumb
 	Thumb string `json:"thumb,omitempty"`
